Project setup and Window creation.
Set up the Csharp project file, including the OpenTk dependency. Chosen
.NET Framework is the latest one available at this moment, 4.8, so be
sure to install it in your development machine.

The program itself right now is just opening a window using OpenTk
framework. This creates the needed OpenGL context, although no rendering
is performed right now.

The window can be closed either with the mouse, clicking on the close
window icon, or with the ESC key. This is just an example and may be
modified as desired in the future.

The program presents a structure and some notes in comments suggesting
to define an interface for the application layer. This is also something
to consider in the future.