Define application interface
A Sharpen based program needs to implement the IApplication interface,
defined in Sharpen.Interface. Application behavior is defined through
this implementation.

The main driver for the visuals is still the Window class, now in
Sharpen.RenderEngine namespace. The creation of the Window now needs a
reference to an IApplication and the main program needs to execute the
Window's Run method.

A basic structure for a Sharpen based program would then be:

    * A class to model the main program. This class is expected to
    create an IApplication object and the Window object. The main
    function will end up giving control to Sharpen by calling Window.Run
    method.
    * A class implementing the IApplication interface. At this moment,
    it's needed just three methods
        - Engage, which will be called once the OpenGL context is ready.
        Here, initializations of the IApplication are needed.
        - Step, to include the application logic during runtime. It is
        executed by the Window object at the frequency indicated at
        its construction.
        - IsFinished, returning a boolean indicating if the application
        needs to finish.

Pending: The implementation of the IApplication, TestApplication, is
currently using OpenTk to read the keyboard. This should not be this
way, as Sharpen is expecting to abstract the other technologies being
used.