Basic render system
The basic render mechanics is in charge of drawing a simple
static model. The model is in 3D coordinates but at this moment
it is actually 2D, being the Z coordinate always 0.

Basic render system is made of three classes:
    - BasicRenderer, which performs the drawing operations;
    - Mesh, which represents a 3D model mesh;
    - Loader, which is in charge of creating the meshes from
      model data and loading them to the OpenGL pipeline.

Each of this clases is having an object in runtime. At this
moment, the Window is owning and creating these objects.

During runtime, the OnRenderFrame method of the Window class
will use the BasicRenderer to draw the frame.

Drawing a frame consists on two steps, each one implemented on
a method from the BasicRenderer class:
    - PrepareFrame, which is clearing the OpenGL color buffer
      from previous frame, using a constant color (right now a
      dark blue color to make it different from the default black);
    - RenderFrame, which performs the OpenGL calls to draw the
      frame: bind the vertex array, enable and draw it.

The OpenGL data arrays are actually created by the Loader class
through its LoadMesh method, which will take the 3D model and
create, and return, a Mesh representing it while creating the
VAO and VBO needed.

Remembering from OpenGL, VAO stands for Vertex Attribute Object
and defines up to 15 different 3D object attributes. Attributes
is an abstract concept and can be anyway needed by the rendering
pipeline: vertices coordinates, indexes, texture coordinates, etc.
At this moment, only the vertices positions are used.

VBO stands for Vertex Buffer Object and is used to contain the
values for each of the attributes. For instance, the first
attribute on the VAO is meant to hold the vertices positions; this
implies as many values as 3x the number of vertices. We need an
array for that and the VBO is this array. The VBO will thus
contain all the positions.