Create meshes at application level
Previous version was creating the example mesh inside the Window
class. This is clearly not desired, as meshes belong to the
application domain. Moving that was not so obvious, as the
renderer needs to access the meshes, and they are actually
created by the Loader anyway, both clases belonging to Sharpen
domain.

To decouple both application and Sharpen worlds, a new static
class has been created to represent the engine. This class is
named Engine and allows creating and running the window (thus,
now the window object belongs to Engine, not to application),
holds the loader object and maintains a list of meshes.

The application now needs to create the window (CreateWindow),
run it (Run) and get the loader (Engine.Loader()). The rest
is unchanged. With this approach, the application can easily
define the meshes and load them using the loader:

    Mesh myMesh = Engine.Loader().LoadMesh(vertices);