Classes visibility update
In C# classes can be public. This is needed to allow visibility
outside the assembly. By default, classes are declared as internal,
meaning that they will be visible only in the same assembly.

As in the future, probably we will be having additional assemblies,
it's good to start defining proper visiblity on the classes.