Render using indices of vertex
If you check the model, you will see that the same vertex is defined
several times, as it is shared by different triangles. This is
inefficient, as you will have 3 floats repeated in the vertex list.

Instead, OpenGL allows defining the vertices from one side and a
list of indices of the vertex used to build each of the triangles.

So, for instance, you can have the vertexes:
    -0.5f,  0.5f, 0f
    -0.5f, -0.5f, 0f
     0.5f,  0.5f, 0f

And define a triangle by means of the indices of these vertex
    0, 1, 2

For a single triangle there is no benefit, but when using a model
of hundreds of vertices, this can make a difference.