Set proper FPS and fix Engine.Run parameters order
Currently, the FPS is set to "as much as possible". This is causing
different frame rates. It has been seen frame rates of 75 or 54.

To make it more consistent, this has been changed on the Test to
60 FPS and now it is steady to 60 with sporadic 59, which is more
than acceptable.

Besides, it's been detected that the parameters of Engine.Run where
order-swapped. Now they will be in the correct order.