Add shaders support and use of IDispose interface
Shaders are supported by the new class ShaderProgram. It searches
shaders written in glsl, compile them and creates a shader program
ready to be run. The source code are to be located in a supplied
directory and the names are hardcoded to "vertex.glsl" and
"fragment.glsl". At this moment, only these stages are supported
but additional ones can be easily added by supplying its details
in the ShaderProgram constructor.

The BasicRenderer class is modified to use shaders. A constructor
will now create the shader program named "Basic" (which will cause
to look for glsl files inside Shaders/Basic directory). The render
process now starts the shader program, draws and stops the shader.

Additional shader programs can be created if the needed glsl files
are located in different directories, but at this moment, only
basic shader is supplied.

Also, now the IDispose interface is being used to release OpenGL
resources when quitting.