Fix detection of shader compilation failure
Condition to detect a failure compilation of any of the shaders
is checking the flag of compilation success for the last compiled
shader, but, if any of the previous shaders failed, it is not
detected by this condition.

Condition now checks the number of successfully compiled shaders
against the total number of expected shaders. If the number differ,
then some of the compilation has failed. The log message will
contain the information of which shader(s) failed and the exception
will be raised.