Add missing method to get attribute location
ShaderProgram is caching in an internal dictionary the location of
all of the attributes used by its shaders. There is currently no
public method supplied to get this information. This fix is adding
that method.

The new method GetAttributeLocation allows getting the position of
the given attribute (the name, a string) for later use.