Add textures and entity support
Texture OpenGL objects are created by the new Texture class. The
created texture will have a static configuration, transparent to
clients. Linear MipMap Linear function will be used as the
minifying filter; Linear will be used for the magnifying filter;
mipmaps are created; and repeat function will be used as the
wrapping method. The pixel data of the texture is loaded into a
newly created texture2D OpenGl object to be passed to the shaders
as sampler.

A new class Entity is encompassing now all the data related to the
model object: the mesh and the used texture. The Loader class still
has the same method to create a Mesh, but now it also has a method
to create a Texture and an Entity. Clients shall use the LoadEntity
method to create one. Refactor is likely to happen at some point
to the Loader class to simplify its public interface. When creating
the Mesh, an additional buffer is linked into the VAO to supply to
the shader program the texture coordinates for each vertex.

BasicRenderer and Window classes have been adapted to use the new
Entity class instead of using directly the Mesh class. Besides,
the renderer now is in charge of starting and stopping its shader
program when asked to render, instead of being done externally,
thoug the public interfaces to start/stop the program are still
present. The Window class is not any longer activating the VAO as
this is done now on the Entity class, through bindToRender public
method (and releaseFromRender to unbind it).

Shaders have been modified to get the color from the texture and
is now receiving as input the texture coordinates to use and the
sampler (the texture itself).

Finally, the test is now defining the texture coordinates for each
of the model vertices and is using the Entity class instead of the
Mesh.