Transformation for position, rotation and scale
Entity can now supply their transformation matrix to change its
position, orientation and scale. The TestGame has now some Logic
to exercise this: it changes the X position and the scale, driven
by the execution time (3 seconds goes towards the left while
enlarging; 3 seconds goes to the right while shrinking -- do not
confuse this enlarging-shrinkng loops with depth management,
although when running it may look as if the entity goes far away).

For the transformation matrix, the Entity has now a method that
supplies it. The BasicRenderer uses this method to pass the matrix
to the ShaderProgram, by means of a new method which will load the
matrix to the vertex shader stage. At the same time, the GLSL code
for the vertex shader is now using the transformation matrix to
calculate each vertex position.

For the time usage, the Engine static class is now supplying a
property named RunningTime which contains the total amount of time
ellapsed since the beggining. This time is updated at the Window
class each time an application loop is executed.