Add Camera
Camera class to control position and orientation of the camera. The
class includes the method to calculate the view matrix to be used
by each Entity.

Shader has been adapted to use the new matrix. Actually, now the
shader recieves a single matrix for the complete model, view and
projection operations.

The test example has been adapted as well to use the camera,
although lack of background is not helping to identify entity
movement from camera movement.

Improvements needed, as the interface is becoming a bit messy.
Each time more stuff goes inside the static class Entity. This
can look bulky to the client application and needs refactoring.