Create solution. Split Sharpen and Test projects.
Now the project is a solution made of Sharpen project, which will
create a dynamic library, and Test project, which will create an
executable using Sharpen library.

This splitting will help in seeing both the application and the
engine as different projects. The engine interface shall be now
clearly exported and the internals shall be not accessible to
the application. This will be changed in next commit.