Visibility modifications
Visibility of classes, methods and attributes have been updated
to restrict access from the applications to internal data. Now
only the really needed data is accessible from application side.

Currently, applications can use:
    - Engine static class.
    - Loader
    - Entity
    - Camera

From the Engine class, the method to get the Basic Renderer object
is now not visible for the application. That method is needed just
internally, so it needs to be there, but the application shall not
be messing around with the renderer.